+++ app/Models/User.php
@@ -74,6 +74,7 @@ class User extends Authenticatable
         return [
             'email_verified_at' => 'datetime',
             'password' => 'hashed',
+            'is_instance_owner' => 'boolean',
         ];
     }
 }